test_real_agent_squad.py
#!/usr/bin/env python3
"""
Teste Real com Agent Squad Framework
Testa a integração completa do FalaChefe Python com Agent Squad.
"""

import asyncio
import json
import sys
import os
from datetime import datetime
from pathlib import Path

# Adiciona o projeto ao path
project_root = Path(__file__).parent
sys.path.insert(0, str(project_root))

# Carrega variáveis de ambiente
from dotenv import load_dotenv
load_dotenv(dotenv_path='.env.python')

# Mock das dependências que não temos
class MockConfig:
    def __init__(self):
        self.openai_api_key = os.getenv('OPENAI_API_KEY', 'test-key')
        self.openai_model = 'gpt-3.5-turbo'
        self.debug = True
        self.log_level = 'INFO'
        self.log_file = 'falachefe_python.log'
        self.api_timeout = 30
        self.typescript_api_url = None

class MockLogger:
    def info(self, msg): print(f"INFO: {msg}")
    def error(self, msg): print(f"ERROR: {msg}")
    def warning(self, msg): print(f"WARNING: {msg}")

def get_component_logger(name):
    return MockLogger()

# Mock dos retrievers
class MockKnowledgeRetriever:
    def __init__(self, config):
        self.config = config
    
    async def retrieve(self, query):
        return [{"content": f"Conhecimento relevante para: {query}", "score": 0.8}]

class LeoKnowledgeRetriever(MockKnowledgeRetriever):
    pass

class MaxKnowledgeRetriever(MockKnowledgeRetriever):
    pass

class LiaKnowledgeRetriever(MockKnowledgeRetriever):
    pass

# Mock do FalaChefe Python
class MockFalaChefePython:
    def __init__(self):
        self.config = MockConfig()
        self.logger = MockLogger()
        
        # Inicializa o Agent Squad
        self.agent_squad = self._initialize_agent_squad()
        self._register_agents()
        
        self.logger.info("FalaChefe Python com Agent Squad inicializado")
    
    def _initialize_agent_squad(self):
        """Inicializa o Agent Squad Framework."""
        try:
            from agent_squad.orchestrator import AgentSquad
            from agent_squad.classifiers import OpenAIClassifier, OpenAIClassifierOptions
            
            # Configura o classificador OpenAI
            classifier_options = OpenAIClassifierOptions(
                api_key=self.config.openai_api_key,
                model_id="gpt-3.5-turbo"
            )
            classifier = OpenAIClassifier(classifier_options)
            
            # Cria o Agent Squad com classificador OpenAI
            return AgentSquad(classifier=classifier)
        except Exception as e:
            self.logger.error(f"Erro ao inicializar Agent Squad: {str(e)}")
            raise
    
    def _register_agents(self):
        """Registra os agentes especializados."""
        try:
            from agent_squad.agents import OpenAIAgent, OpenAIAgentOptions
            
            # Agente Leo (Financeiro)
            leo_options = OpenAIAgentOptions(
                name="leo",
                description="Mentor financeiro experiente e confiável. Ajuda a entender números, evitar erros financeiros e planejar o caixa.",
                model="gpt-3.5-turbo",
                api_key=self.config.openai_api_key,
                inference_config={"temperature": 0.4, "max_tokens": 800}
            )
            leo_agent = OpenAIAgent(leo_options)
            self.agent_squad.add_agent(leo_agent)
            
            # Agente Max (Marketing)
            max_options = OpenAIAgentOptions(
                name="max",
                description="Especialista em marketing e vendas. Ajuda a criar campanhas, gerar leads e aumentar vendas.",
                model="gpt-3.5-turbo",
                api_key=self.config.openai_api_key,
                inference_config={"temperature": 0.6, "max_tokens": 800}
            )
            max_agent = OpenAIAgent(max_options)
            self.agent_squad.add_agent(max_agent)
            
            # Agente Lia (RH)
            lia_options = OpenAIAgentOptions(
                name="lia",
                description="Especialista em recursos humanos. Ajuda com gestão de pessoas, processos de RH e desenvolvimento de equipe.",
                model="gpt-3.5-turbo",
                api_key=self.config.openai_api_key,
                inference_config={"temperature": 0.5, "max_tokens": 800}
            )
            lia_agent = OpenAIAgent(lia_options)
            self.agent_squad.add_agent(lia_agent)
            
            self.logger.info("Agentes registrados: Leo (Financeiro), Max (Marketing), Lia (RH)")
            
        except Exception as e:
            self.logger.error(f"Erro ao registrar agentes: {str(e)}")
            raise
    
    async def process_conversation(self, conversation_data):
        """Processa uma conversa."""
        try:
            message = conversation_data.get('message', '')
            user_id = conversation_data.get('user_id', 'unknown')
            session_id = conversation_data.get('session_id', 'default')
            
            self.logger.info(f"Processando conversa: {conversation_data.get('id', 'unknown')}")
            
            # Processa com Agent Squad
            response = await self.agent_squad.route_request(
                message,
                user_id,
                session_id,
                {},
                True  # streaming
            )
            
            # Extrai resposta
            response_content = ""
            if hasattr(response, 'output'):
                async for chunk in response.output:
                    if hasattr(chunk, 'text'):
                        response_content += chunk.text
            
            agent_name = response.metadata.agent_name if hasattr(response, 'metadata') else 'unknown'
            
            return {
                "response": {
                    "message": response_content,
                    "agent_name": agent_name,
                    "user_id": user_id,
                    "session_id": session_id,
                    "timestamp": datetime.now().isoformat(),
                    "success": True
                },
                "analysis": {
                    "insights": [f"Processado pelo agente {agent_name}"],
                    "sentiment": "positive",
                    "intent": "general_query"
                },
                "processed_by": "agent_squad_python"
            }
            
        except Exception as e:
            self.logger.error(f"Erro ao processar conversa: {str(e)}")
            return {
                "error": str(e),
                "processed_by": "error_fallback"
            }
    
    async def health_check(self):
        """Verifica a saúde do sistema."""
        try:
            return {
                "timestamp": datetime.now().isoformat(),
                "status": "healthy",
                "components": {
                    "agent_squad": {"status": "healthy"},
                    "agents": {
                        "leo": "healthy",
                        "max": "healthy", 
                        "lia": "healthy"
                    }
                }
            }
        except Exception as e:
            return {
                "timestamp": datetime.now().isoformat(),
                "status": "unhealthy",
                "error": str(e)
            }

async def run_real_test():
    """Executa teste real do Agent Squad."""
    print("🚀 INICIANDO TESTE REAL DO AGENT SQUAD FRAMEWORK")
    print("="*60)
    
    try:
        # Inicializa o FalaChefe Python
        print("🔧 Inicializando FalaChefe Python com Agent Squad...")
        falachefe = MockFalaChefePython()
        print("✅ FalaChefe Python inicializado com sucesso!")
        
        # Teste 1: Health Check
        print("\n🏥 Teste 1: Health Check")
        health = await falachefe.health_check()
        print(f"Status: {health['status']}")
        print(f"Componentes: {health['components']}")
        
        # Teste 2: Processamento de conversas
        print("\n💬 Teste 2: Processamento de Conversas")
        
        test_conversations = [
            {
                "id": "test_001",
                "message": "Preciso de ajuda com finanças da minha empresa",
                "user_id": "user_001",
                "session_id": "session_001"
            },
            {
                "id": "test_002",
                "message": "Como criar uma campanha de marketing no WhatsApp?",
                "user_id": "user_002", 
                "session_id": "session_002"
            },
            {
                "id": "test_003",
                "message": "Preciso de ajuda com gestão de funcionários",
                "user_id": "user_003",
                "session_id": "session_003"
            }
        ]
        
        for i, conv in enumerate(test_conversations, 1):
            print(f"\n  Teste {i}: {conv['message'][:50]}...")
            
            try:
                result = await falachefe.process_conversation(conv)
                
                if result.get('error'):
                    print(f"    ❌ Erro: {result['error']}")
                    continue
                
                response = result.get('response', {})
                agent_used = response.get('agent_name', 'unknown')
                message_content = response.get('message', 'N/A')
                
                print(f"    ✅ Agente usado: {agent_used}")
                print(f"    📝 Resposta: {message_content[:100]}...")
                
                # Verifica análise
                analysis = result.get('analysis', {})
                if analysis:
                    print(f"    📊 Insights: {len(analysis.get('insights', []))} itens")
                
            except Exception as e:
                print(f"    ❌ Erro no teste {i}: {str(e)}")
        
        print("\n🎉 TESTE REAL CONCLUÍDO COM SUCESSO!")
        print("✅ Agent Squad Framework funcionando corretamente")
        print("✅ Agentes Leo, Max e Lia registrados")
        print("✅ Processamento de conversas operacional")
        
    except Exception as e:
        print(f"❌ ERRO FATAL: {str(e)}")
        import traceback
        traceback.print_exc()

if __name__ == "__main__":
    asyncio.run(run_real_test())
